Menu/Filters translations now save in the same transaction as the settings document
