--- app/api/core/infrastructure/factories/UpdateTranslationContextUseCaseFactory.ts
@@ -1,14 +1,0 @@
-import { UpdateTranslationContextUseCase } from '#api/core/application/UpdateTranslationContext.js';
-import { TranslationsServiceFactory } from './TranslationsServiceFactory.js';
-import { TransactionManagerFactory } from './TransactionManagerFactory.js';
-
-export class UpdateTranslationContextUseCaseFactory {
-  static default() {
-    const transactionManager = TransactionManagerFactory.default();
-
-    return new UpdateTranslationContextUseCase({
-      transactionManager,
-      translationsService: TranslationsServiceFactory.default({ transactionManager }),
-    });
-  }
-}
