+++ app/api/settings/settings.ts
@@ -1,123 +1,16 @@
-import {
-  Settings,
-  SettingsLinkSchema,
-  SettingsFilterSchema,
-  SettingsSublinkSchema,
-} from '#shared/types/settingsType.js';
+import { Settings, SettingsFilterSchema } from '#shared/types/settingsType.js';
 import { ensure } from '#shared/tsUtils.js';
 import { LanguageSchema, LatLonSchema, ObjectIdSchema } from '#shared/types/commonTypes.js';
 
 import { validateSettings } from '#shared/types/settingsSchema.js';
 import { ArrayUtils } from '#api/common.v2/utils/Array.js';
 import { TemplateFacade } from '#api/core/infrastructure/facades/TemplateFacade.js';
-import { UpdateTranslationContextUseCaseFactory } from '#api/core/infrastructure/factories/UpdateTranslationContextUseCaseFactory.js';
-import { settingsModel } from './settingsModel.js';
 import { TemplatesDAOFactory } from '#api/core/infrastructure/factories/TemplatesDAOFactory.js';
 import { TemplateDBO } from '#api/core/infrastructure/mongodb/template/DBOs/TemplateDBO.js';
+import { settingsModel } from './settingsModel.js';
+import { persistSettingsAndTranslations } from './settingsTranslations.js';
 
 const DEFAULT_MAP_STARTING_POINT: LatLonSchema[] = [{ lon: 6, lat: 46 }];
-
-type FilterOrLink = SettingsFilterSchema | SettingsLinkSchema | SettingsSublinkSchema;
-
-const isLink = (item: any): item is SettingsLinkSchema => item.type && item.title;
-
-const getUpdatesAndDeletes = <T extends FilterOrLink>(
-  matchProperty: keyof T,
-  propertyName: keyof T,
-  newValues: T[] = [],
-  currentValues: T[] = []
-) => {
-  const updatedValues: { [k: string]: any } = {};
-  const deletedValues: string[] = [];
-  const values: { [k: string]: string } = {};
-
-  //flatten values
-  const flattenedCurrentValues = currentValues.reduce((result, value) => {
-    if (isLink(value) && value.sublinks) {
-      return [...result, ...(value.sublinks as T[]), value];
-    }
-    return [...result, value];
-  }, [] as T[]);
-
-  const flattenedNewValues = newValues.reduce<T[]>((result, value) => {
-    if (isLink(value) && value.sublinks) {
-      return [...result, ...(value.sublinks as T[]), value];
-    }
-    return [...result, value];
-  }, [] as T[]);
-
-  flattenedCurrentValues.forEach(value => {
-    const matchValue = flattenedNewValues.find(
-      (v): v is T =>
-        v[matchProperty] && v[matchProperty]?.toString() === value[matchProperty]?.toString()
-    );
-
-    if (!matchValue) {
-      deletedValues.push(ensure<string>(value[propertyName]));
-      return;
-    }
-
-    const nameHasChanged = value[propertyName] !== matchValue[propertyName];
-    if (nameHasChanged) {
-      updatedValues[ensure<string>(value[propertyName])] = matchValue[propertyName];
-    }
-  });
-
-  //latest values
-  flattenedNewValues.forEach(value => {
-    values[ensure<string>(value[propertyName])] = ensure<string>(value[propertyName]);
-  });
-
-  return { updatedValues, deletedValues, values };
-};
-
-const saveLinksTranslations = async (
-  newLinks: Settings['links'],
-  currentLinks: Settings['links'] = []
-) => {
-  if (!newLinks) {
-    return Promise.resolve();
-  }
-
-  const { updatedValues, deletedValues, values } = getUpdatesAndDeletes(
-    '_id',
-    'title',
-    newLinks,
-    currentLinks
-  );
-
-  return UpdateTranslationContextUseCaseFactory.default().execute({
-    context: { id: 'Menu', label: 'Menu', type: 'Uwazi UI' },
-    keyChanges: updatedValues,
-    keysToDelete: deletedValues,
-    valueChanges: values,
-  });
-};
-
-const saveFiltersTranslations = async (
-  _newFilters: Settings['filters'],
-  _currentFilters: Settings['filters'] = []
-) => {
-  if (!_newFilters) {
-    return Promise.resolve();
-  }
-
-  const newFilters = _newFilters.filter(item => item.items);
-  const currentFilters = _currentFilters.filter(item => item.items);
-
-  const { updatedValues, deletedValues, values } = getUpdatesAndDeletes(
-    'id',
-    'name',
-    newFilters,
-    currentFilters
-  );
-  return UpdateTranslationContextUseCaseFactory.default().execute({
-    context: { id: 'Filters', label: 'Filters', type: 'Uwazi UI' },
-    keyChanges: updatedValues,
-    keysToDelete: deletedValues,
-    valueChanges: values,
-  });
-};
 
 function removeTemplate(filters: SettingsFilterSchema[], templateId: ObjectIdSchema) {
   const filterTemplate = (filter: SettingsFilterSchema) => filter.id !== templateId;
@@ -142,6 +35,16 @@ function setDefaults(storedSettings: Settings[]) {
   return settings;
 }
 
+async function updateTemplatesForNewNameGeneration(currentSettings: Settings) {
+  const dao = TemplatesDAOFactory.default();
+  const templates = (await dao.get()) as TemplateDBO[];
+  const defaultLanguage = currentSettings?.languages?.find(l => l.default)?.key!;
+
+  await ArrayUtils.sequentialFor(templates, async (template: any) => {
+    await TemplateFacade.update({ ...template, reindex: false }, defaultLanguage);
+  });
+}
+
 export default {
   async get(query: any = {}, select: any = '') {
     return ensure<Settings>(
@@ -152,19 +55,10 @@ export default {
   async save(settings: Settings) {
     await validateSettings(settings);
     const currentSettings = await this.get();
-    await saveLinksTranslations(settings.links, currentSettings.links);
-    await saveFiltersTranslations(settings.filters, currentSettings.filters);
-
-    const result = await settingsModel.save({ ...settings, _id: currentSettings._id });
+    const result = await persistSettingsAndTranslations(settings, currentSettings);
 
     if (!currentSettings.newNameGeneration && settings.newNameGeneration) {
-      const dao = TemplatesDAOFactory.default();
-      const templates = (await dao.get()) as TemplateDBO[];
-      const defaultLanguage = currentSettings?.languages?.find(l => l.default)?.key!;
-
-      await ArrayUtils.sequentialFor(templates, async (template: any) => {
-        await TemplateFacade.update({ ...template, reindex: false }, defaultLanguage);
-      });
+      await updateTemplatesForNewNameGeneration(currentSettings);
     }
 
     return result;
@@ -207,34 +101,34 @@ export default {
   },
 
   async removeTemplateFromFilters(templateId: ObjectIdSchema) {
-    const settings = await this.get();
+    const currentSettings = await this.get();
 
-    if (!settings.filters) {
+    if (!currentSettings.filters) {
       return Promise.resolve();
     }
 
-    settings.filters = removeTemplate(settings.filters, templateId);
-    return this.save(settings);
+    currentSettings.filters = removeTemplate(currentSettings.filters, templateId);
+    return this.save(currentSettings);
   },
 
   async updateFilterName(filterId: ObjectIdSchema, name: string) {
-    const settings = await this.get();
+    const currentSettings = await this.get();
 
-    if (!(settings.filters || []).some(eachFilter => eachFilter.id === filterId)) {
+    if (!(currentSettings.filters || []).some(eachFilter => eachFilter.id === filterId)) {
       return Promise.resolve();
     }
 
-    const filter = (settings.filters || []).find(eachFilter => eachFilter.id === filterId);
+    const filter = (currentSettings.filters || []).find(eachFilter => eachFilter.id === filterId);
     if (filter) {
       filter.name = name;
     }
 
-    return this.save(settings);
+    return this.save(currentSettings);
   },
 
   async getLinks() {
-    const settings = await this.get();
-    return settings.links || [];
+    const currentSettings = await this.get();
+    return currentSettings.links || [];
   },
 
   async saveLinks(links: Settings['links']) {
